Login/Subscribe page responsive, styles files reorganized in one styles folder with pages and components folders in it, clean css on desktop to be more responsive for mobile in progress
