+++ frontend/src/components/HeaderBannerHorses.jsx
