+++ frontend/src/components/LoansList.jsx
@@ -29,15 +29,17 @@ function LoanList({
   };
 
   return (
-    <div className="loanList">
-      <li className="dash_vehicle_elements_list">
-        <div className="dash_vehicle_elements_list_image">
-          <img src={pathToImages} alt="profile of our horse" />
-        </div>
-        <div className="dash_vehicle_elements_list_paragraph description">
-          <div>
-            <p className="boldText">{model}</p>
-            <p>{type}</p>
+    <div>
+      <li className="vehiclesList">
+        <div className="pictureContent">
+          <div className="dash_vehicle_elements_list_image">
+            <img src={pathToImages} alt="profile of our horse" />
+          </div>
+          <div className="description">
+            <div>
+              <p className="boldText">{model}</p>
+              <p>{type}</p>
+            </div>
           </div>
         </div>
         <div>
