+++ frontend/src/styles/components/chart.scss
