+++ frontend/src/styles/components/headerBannerHorses.scss
