+++ frontend/src/styles/components/horseCard.scss
