+++ frontend/src/styles/components/loanList.scss
@@ -1,0 +1,31 @@
+  .vehiclesList {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 30px;
+    gap: 30px;
+    width: 100%;
+    height: 140px;
+    margin-bottom: 30px;
+    background-color: white;
+    border-radius: 10px;
+    box-shadow: $shadowCard;
+    height: auto;
+
+    .pictureContent {
+      display: flex;
+      flex-direction: row;
+      gap: 30px;
+      align-items: center;
+    }
+
+    .description {
+      display: flex;
+
+      div {
+        display: flex;
+        flex-direction: column;
+        gap: 8px;
+      }
+    }
+  }
