+++ frontend/src/styles/components/ourVehicles.scss
