+++ frontend/src/styles/components/presentation.scss
