+++ frontend/src/styles/components/sideBarMenu.scss
@@ -2,13 +2,9 @@
   display: flex;
   flex-direction: column;
   justify-content: space-between;
-  min-width: 20vw;
+  min-width: 360px;
   height: 100%;
   background-color: white;
-  position: fixed;
-  z-index: 1;
-  left: 0;
-  top: 0;
   box-shadow: $shadowCard;
   padding: 60px;
   gap: 40px;
