+++ frontend/src/styles/components/visualLogin.scss
